[sample] address rows through an int64 offset
row * stride_b is an int32 multiply: Triton binds a stride that fits in int32
as int32, and the row index is int32 too, so the product wraps once the tensor
passes 2**31 elements -- 13662 rows at vocab 157184, which a wide enough
rollout reaches.

The wrap is not a wrong answer. The offset addresses outside the tensor and the
device comes down with a vector core exception, taking the rest of the process
with it. On 910B3, bf16, vocab 157184: 13262 rows matched before and after,
13918 rows raised runtime result = 507035 before and matches after.

The multiply runs once per row, outside the tile loop, so the cost is not
measurable.

What trips it is the element count, so the smallest case that reaches the
boundary allocates ~4 GiB of logits. The test skips itself where that does not
fit and sits last in the file: a regression here fails the session, not just
its own case.

+++ python/sgl_kernel_npu/sgl_kernel_npu/sample/argmax_softmax_prob.py
@@ -41,7 +41,11 @@ def _argmax_prob_kernel(
     row_pid = tl.program_id(0)
     n_prog = tl.num_programs(0)
     for row in tl.range(row_pid, B, n_prog):
-        base = row * stride_b
+        # int64 before the multiply: Triton binds a stride that fits in int32 as
+        # int32, so row * stride_b wraps once the tensor passes 2**31 elements
+        # (13662 rows at vocab 157184) and addresses outside it. Once per row,
+        # not in the tile loop below.
+        base = row.to(tl.int64) * stride_b
         m = tl.full((), float("-inf"), tl.float32)
         arg = tl.zeros((), tl.int64)
         s = tl.zeros((), tl.float32)
